+++ lib/src/configuration/better_player_drm_type.dart
@@ -1,4 +1,4 @@
 ///Types of available DRM's
 enum BetterPlayerDrmType{
-  AES
+  TOKEN
 }