+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * Disabled picture in picture in fullscreen mode.
 * Fixed enabled parameter for skip back and forward.
 * Fixed notification configuration null issue (by https://github.com/bounty1342)
-* Added AES DRM support.
+* Added token based DRM support.
 * Updated documentation.
 
 ## 0.0.57